requirements.txt
Django>=5.2.9
celery[redis]
chardet
cryptography
django-axes
django-celery-beat
django-celery-results
django-cors-headers>=4.3.0
django-simple-captcha
django-vite
django-waffle==5.0.0
djangorestframework==3.16.1
gevent>=1.4
gunicorn
numpy==2.3.4
pandas==2.3.3
Pillow
psycopg2-binary
python-dateutil==2.9.0.post0
python-dotenv
pytz==2025.2
requests>=2.25.1
six==1.17.0
tzdata==2025.2
whitenoise
scipy
playwright>=1.40.0

# Energy Revenue Hub: PDF text (pdfplumber/PyPDF2) + OCR for scanned PDFs (Celery worker needs these too)
pdfplumber>=0.11.0
PyPDF2>=3.0.0
pymupdf>=1.24.0
easyocr>=1.7.0
reportlab>=4.0.0
weasyprint>=62.0

# Engineering Tools: Solar layout KML/KMZ export (geospatial)
geopandas>=0.14.0
shapely>=2.0.0
simplekml>=1.3.6
pyproj>=3.6.0